web/requirements.txt
flask
flask-dance[sqla]
sqlalchemy
tweepy
python-dotenv
gunicorn